feat: moved svgs to public assets folder
