+++ src/components/Header.tsx
@@ -4,7 +4,7 @@ import NextLink from 'next/link';
 import { signOut } from 'next-auth/react';
 import { FC, useState } from 'react';
 
-import MumbleLogo from '../assets/svg/mumbleLogo.svg';
+import MumbleLogo from '../../public/assets/svg/mumbleLogo.svg';
 import { TUser } from '../types';
 import UserSettings from './form/UserSettings';
 import { UserProfile } from './user/UserProfile';
